adjust testValidators  ElectionRoundFormTemplateAssignment #4
