+++ api/src/Feature.FormTemplates/Create/Endpoint.cs
@@ -27,7 +27,7 @@ public class Endpoint(IRepository<FormTemplate> repository) :
             .ToList()
             .AsReadOnly();
         
-        var formTemplate = Vote.Monitor.Domain.Entities.FormTemplateAggregate.FormTemplate.Create(req.FormType, req.Code, req.DefaultLanguage, req.Name, req.Description, req.Languages, questions);
+        var formTemplate = FormTemplate.Create(req.FormType, req.Code, req.DefaultLanguage, req.Name, req.Description, req.Languages, questions);
 
         await repository.AddAsync(formTemplate, ct);
