+++ api/src/Vote.Monitor.Api/appsettings.Development.json
@@ -1,48 +1,48 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Information",
-      "Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware": "None"
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft.AspNetCore": "Information",
+            "Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware": "None"
+        }
+    },
+    "Domain": {
+        "DbConnectionConfig": {
+            "Server": "localhost",
+            "Port": "5432",
+            "Database": "vote-monitor",
+            "UserId": "postgres",
+            "Password": "docker"
+        }
+    },
+    "Core": {
+        "EnableHangfire": true,
+        "HangfireConnectionConfig": {
+            "Server": "localhost",
+            "Port": "5432",
+            "Database": "vote-monitor-hangfire",
+            "UserId": "postgres",
+            "Password": "docker"
+        }
+    },
+    "Sentry": {
+        "Enabled": false,
+        "Dsn": "",
+        "TracesSampleRate": 0.2
+    },
+    "ApiConfiguration": {
+        "WebAppUrl": "http://localhost:5173/"
+    },
+    "Seeders": {
+        "PlatformAdminSeeder": {
+            "FirstName": "John",
+            "LastName": "Doe",
+            "Email": "john.doe@example.com",
+            "PhoneNumber": "string",
+            "Password": "string"
+        }
+    },
+    "Statistics": {
+        "ApiKey": "dummy-api-key"
     }
-  },
-  "Domain": {
-    "DbConnectionConfig": {
-      "Server": "localhost",
-      "Port": "5432",
-      "Database": "vote-monitor",
-      "UserId": "postgres",
-      "Password": "docker"
-    }
-  },
-  "Core": {
-    "EnableHangfire": true,
-    "HangfireConnectionConfig": {
-      "Server": "localhost",
-      "Port": "5432",
-      "Database": "vote-monitor-hangfire",
-      "UserId": "postgres",
-      "Password": "docker"
-    }
-  },
-  "Sentry": {
-    "Enabled": false,
-    "Dsn": "",
-    "TracesSampleRate": 0.2
-  },
-  "ApiConfiguration": {
-    "WebAppUrl": "http://localhost:5173/"
-  },
-  "Seeders": {
-    "PlatformAdminSeeder": {
-      "FirstName": "John",
-      "LastName": "Doe",
-      "Email": "john.doe@example.com",
-      "PhoneNumber": "string",
-      "Password": "string"
-    }
-  },
-  "Statistics": {
-    "ApiKey": "dummy-api-key"
-  }
 }
