+++ api/src/Vote.Monitor.Hangfire/appsettings.Development.json
@@ -1,72 +1,72 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Information",
-      "Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware": "None",
-      "Hangfire": "Information"
-    }
-  },
-  "Serilog": {
-    "MinimumLevel": {
-      "Default": "Information",
-      "Override": {
-        "Microsoft": "Warning",
-        "System": "Warning"
-      }
-    }
-  },
-  "Domain": {
-    "DbConnectionConfig": {
-      "Server": "localhost",
-      "Port": "5432",
-      "Database": "vote-monitor",
-      "UserId": "postgres",
-      "Password": "docker"
-    }
-  },
-  "Core": {
-    "HangfireConnectionConfig": {
-      "Server": "localhost",
-      "Port": "5432",
-      "Database": "vote-monitor-hangfire",
-      "UserId": "postgres",
-      "Password": "docker"
-    }
-  },
-  "Mailing": {
-    "MailSenderType": "SMTP",
-    "SMTP": {
-      "SenderName": "John Doe",
-      "SenderEmail": "john@example.com",
-      "Host": "localhost",
-      "Port": 1025,
-      "UserName": "user@example.com",
-      "Password": "password123"
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft.AspNetCore": "Information",
+            "Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware": "None",
+            "Hangfire": "Information"
+        }
     },
-    "SES": {
-      "SenderName": "John Doe",
-      "SenderEmail": "john@example.com",
-      "AWSRegion": "eu-central-1",
-      "AWSAccessKey": "youraccesskey",
-      "AWSSecretKey": "yoursecretkey"
-    }
-  },
-  "FileStorage": {
-    "FileStorageType": "LocalDisk",
-    "LocalDisk": {
-      "Path": "Uploads"
+    "Serilog": {
+        "MinimumLevel": {
+            "Default": "Information",
+            "Override": {
+                "Microsoft": "Warning",
+                "System": "Warning"
+            }
+        }
     },
-    "S3": {
-      "BucketName": "",
-      "PresignedUrlValidityInSeconds": 3600,
-      "AWSRegion": "eu-central-1",
-      "AWSAccessKey": "youraccesskey",
-      "AWSSecretKey": "yoursecretkey"
+    "Domain": {
+        "DbConnectionConfig": {
+            "Server": "localhost",
+            "Port": "5432",
+            "Database": "vote-monitor",
+            "UserId": "postgres",
+            "Password": "docker"
+        }
+    },
+    "Core": {
+        "HangfireConnectionConfig": {
+            "Server": "localhost",
+            "Port": "5432",
+            "Database": "vote-monitor-hangfire",
+            "UserId": "postgres",
+            "Password": "docker"
+        }
+    },
+    "Mailing": {
+        "MailSenderType": "SMTP",
+        "SMTP": {
+            "SenderName": "John Doe",
+            "SenderEmail": "john@example.com",
+            "Host": "localhost",
+            "Port": 1025,
+            "UserName": "user@example.com",
+            "Password": "password123"
+        },
+        "SES": {
+            "SenderName": "John Doe",
+            "SenderEmail": "john@example.com",
+            "AWSRegion": "eu-central-1",
+            "AWSAccessKey": "youraccesskey",
+            "AWSSecretKey": "yoursecretkey"
+        }
+    },
+    "FileStorage": {
+        "FileStorageType": "LocalDisk",
+        "LocalDisk": {
+            "Path": "Uploads"
+        },
+        "S3": {
+            "BucketName": "",
+            "PresignedUrlValidityInSeconds": 3600,
+            "AWSRegion": "eu-central-1",
+            "AWSAccessKey": "youraccesskey",
+            "AWSSecretKey": "yoursecretkey"
+        }
+    },
+    "DashboardAuth": {
+        "Username": "admin",
+        "Password": "admin"
     }
-  },
-  "DashboardAuth": {
-    "Username": "admin",
-    "Password": "admin"
-  }
 }
