+++ api/src/Vote.Monitor.Hangfire/appsettings.json
@@ -1,84 +1,84 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning",
-      "Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware": "None",
-      "Hangfire": "Information"
-    }
-  },
-  "Serilog": {
-    "MinimumLevel": {
-      "Default": "Information",
-      "Override": {
-        "Microsoft": "Warning",
-        "System": "Warning"
-      }
-    }
-  },
-  "Domain": {
-    "DbConnectionConfig": {
-      "Server": "localhost",
-      "Port": "5432",
-      "Database": "vote-monitor",
-      "UserId": "postgres",
-      "Password": "docker"
-    }
-  },
-  "Core": {
-    "HangfireConnectionConfig": {
-      "Server": "localhost",
-      "Port": "5432",
-      "Database": "vote-monitor-hangfire",
-      "UserId": "postgres",
-      "Password": "docker"
-    }
-  },
-  "Mailing": {
-    "MailSenderType": "SMTP",
-    "SMTP": {
-      "SenderName": "John Doe",
-      "SenderEmail": "john@example.com",
-      "Host": "smtp.example.com",
-      "Port": 587,
-      "UserName": "user@example.com",
-      "Password": "password123"
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft.AspNetCore": "Warning",
+            "Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware": "None",
+            "Hangfire": "Information"
+        }
     },
-    "SES": {
-      "SenderName": "John Doe",
-      "SenderEmail": "john@example.com",
-      "AWSRegion": "eu-central-1",
-      "AWSAccessKey": "youraccesskey",
-      "AWSSecretKey": "yoursecretkey"
-    }
-  },
-  "FileStorage": {
-    "FileStorageType": "LocalDisk",
-    "LocalDisk": {
-      "Path": "Uploads"
+    "Serilog": {
+        "MinimumLevel": {
+            "Default": "Information",
+            "Override": {
+                "Microsoft": "Warning",
+                "System": "Warning"
+            }
+        }
     },
-    "S3": {
-      "BucketName": "",
-      "PresignedUrlValidityInSeconds": 432000,
-      "AWSRegion": "eu-central-1",
-      "AWSAccessKey": "youraccesskey",
-      "AWSSecretKey": "yoursecretkey"
-    }
-  },
-  "PushNotifications": {
-    "SenderType": "Expo",
-    "Firebase": {
-      "Token": "TBD",
-      "BatchSize": 100
+    "Domain": {
+        "DbConnectionConfig": {
+            "Server": "localhost",
+            "Port": "5432",
+            "Database": "vote-monitor",
+            "UserId": "postgres",
+            "Password": "docker"
+        }
     },
-    "Expo": {
-      "ChannelId": "default",
-      "Priority": "high",
-      "TtlSeconds": 259200,
-      "BatchSize": 100
-    }
-  },
-  "DashboardAuth": {
+    "Core": {
+        "HangfireConnectionConfig": {
+            "Server": "localhost",
+            "Port": "5432",
+            "Database": "vote-monitor-hangfire",
+            "UserId": "postgres",
+            "Password": "docker"
+        }
+    },
+    "Mailing": {
+        "MailSenderType": "SMTP",
+        "SMTP": {
+            "SenderName": "John Doe",
+            "SenderEmail": "john@example.com",
+            "Host": "smtp.example.com",
+            "Port": 587,
+            "UserName": "user@example.com",
+            "Password": "password123"
+        },
+        "SES": {
+            "SenderName": "John Doe",
+            "SenderEmail": "john@example.com",
+            "AWSRegion": "eu-central-1",
+            "AWSAccessKey": "youraccesskey",
+            "AWSSecretKey": "yoursecretkey"
+        }
+    },
+    "FileStorage": {
+        "FileStorageType": "LocalDisk",
+        "LocalDisk": {
+            "Path": "Uploads"
+        },
+        "S3": {
+            "BucketName": "",
+            "PresignedUrlValidityInSeconds": 432000,
+            "AWSRegion": "eu-central-1",
+            "AWSAccessKey": "youraccesskey",
+            "AWSSecretKey": "yoursecretkey"
+        }
+    },
+    "PushNotifications": {
+        "SenderType": "Expo",
+        "Firebase": {
+            "Token": "TBD",
+            "BatchSize": 100
+        },
+        "Expo": {
+            "ChannelId": "default",
+            "Priority": "high",
+            "TtlSeconds": 259200,
+            "BatchSize": 100
+        }
+    },
+    "DashboardAuth": {
 
-  }
+    }
 }
